Move away HTTP requests utilities from general.py module
This change improves code organization by grouping utility functions
together rather than having them scattered across different modules.
Moving them to core/utils.py better reflects their purpose and makes
them more easily accessible to other parts of the codebase that need
download functionality.

Furthermore, having general.py module and generic.py directory in
package managers might be a little bit confusing.
