+++ hermeto/core/package_managers/npm.py
@@ -17,9 +17,9 @@ from hermeto.core.models.input import Request
 from hermeto.core.models.output import ProjectFile, RequestOutput
 from hermeto.core.models.property_semantics import PropertySet
 from hermeto.core.models.sbom import Component
-from hermeto.core.package_managers.general import async_download_files
 from hermeto.core.rooted_path import RootedPath
 from hermeto.core.scm import RepoID, clone_as_tarball, get_repo_id
+from hermeto.core.utils import async_download_files
 
 DEPENDENCY_TYPES = (
     "dependencies",
